:recycle: Updated typescript definitions. Moved some interface types to remove reliance from design system on editor-ui user and validation types.
