+++ packages/design-system/src/components/N8nAvatar/index.ts
