+++ packages/design-system/src/components/N8nIcon/index.ts
