+++ packages/design-system/src/components/N8nIconButton/index.ts
