+++ packages/design-system/src/components/N8nSpinner/index.ts
