+++ packages/design-system/src/components/N8nTabs/index.ts
