+++ packages/design-system/src/components/N8nText/index.ts
