+++ packages/design-system/src/components/N8nTooltip/index.ts
