+++ packages/design-system/src/components/N8nUserInfo/index.ts
