+++ packages/design-system/src/components/N8nUsersList/UsersList.vue
@@ -20,7 +20,7 @@
 </template>
 
 <script lang="ts">
-import { IUser } from '../../Interface';
+import { IUser } from '../../types';
 import Vue from 'vue';
 import N8nActionToggle from '../N8nActionToggle';
 import N8nBadge from '../N8nBadge';
