+++ packages/design-system/src/components/ResizeObserver/index.ts
