+++ packages/design-system/src/main.d.ts
@@ -1,0 +1,16 @@
+import Vue from 'vue';
+import * as locale from './locale';
+
+declare module 'vue/types/vue' {
+	interface Vue {
+		$style: Record<string, string>;
+	}
+}
+
+declare module 'n8n-design-system' {
+	export * from './components';
+	export { N8nUserSelect, N8nUsersList } from './components'; // Workaround for circular imports, will be removed when migrated to typescript
+	export { locale };
+}
+
+export * from './types';
