+++ packages/design-system/src/mixins/index.ts
@@ -1,0 +1,1 @@
+export { default as localeMixin } from './locale';
