+++ packages/design-system/src/mixins/locale.ts
@@ -2,7 +2,7 @@ import { t } from '../locale';
 
 export default {
 	methods: {
-		t(...args) {
+		t(...args: string[]) {
 			return t.apply(this, args);
 		},
 	},
