+++ packages/design-system/src/types/index.ts
@@ -1,0 +1,2 @@
+export * from './form';
+export * from './user';
