+++ packages/editor-ui/src/Interface.ts
@@ -22,6 +22,12 @@ import {
 	IWorkflowSettings as IWorkflowSettingsWorkflow,
 	WorkflowExecuteMode,
 } from 'n8n-workflow';
+import {
+	IRole,
+	IUser,
+} from 'n8n-design-system';
+
+export * from 'n8n-design-system/src/types';
 
 declare module 'jsplumb' {
 	interface PaintStyle {
@@ -483,28 +489,6 @@ export interface IVersionNotificationSettings {
 	infoUrl: string;
 }
 
-export type IPersonalizationSurveyAnswersV1 = {
-	codingSkill?: string | null;
-	companyIndustry?: string[] | null;
-	companySize?: string | null;
-	otherCompanyIndustry?: string | null;
-	otherWorkArea?: string | null;
-	workArea?: string[] | string | null;
-};
-
-export type IPersonalizationSurveyAnswersV2 = {
-	version: 'v2';
-	automationGoal?: string | null;
-	codingSkill?: string | null;
-	companyIndustryExtended?: string[] | null;
-	companySize?: string | null;
-	companyType?: string | null;
-	mspFocus?: string[] | null;
-	mspFocusOther?: string | null;
-	otherAutomationGoal?: string | null;
-	otherCompanyIndustryExtended?: string[] | null;
-};
-
 export interface IN8nPrompts {
 	message: string;
 	title: string;
@@ -888,21 +872,6 @@ export interface IBounds {
 
 export type ILogInStatus = 'LoggedIn' | 'LoggedOut';
 
-export type IRole = 'default' | 'owner' | 'member';
-
-export interface IUserResponse {
-	id: string;
-	firstName?: string;
-	lastName?: string;
-	email?: string;
-	globalRole?: {
-		name: IRole;
-		id: string;
-	};
-	personalizationAnswers?: IPersonalizationSurveyAnswersV1 | IPersonalizationSurveyAnswersV2 | null;
-	isPending: boolean;
-}
-
 export interface IInviteResponse {
 	user: {
 		id: string;
@@ -910,59 +879,6 @@ export interface IInviteResponse {
 	};
 	error?: string;
 }
-
-export interface IUser extends IUserResponse {
-	isDefaultUser: boolean;
-	isPendingUser: boolean;
-	isOwner: boolean;
-	fullName?: string;
-}
-
-export type Rule = { name: string; config?: any}; // tslint:disable-line:no-any
-
-export type RuleGroup = {
-	rules: Array<Rule | RuleGroup>;
-	defaultError?: {messageKey: string, options?: any}; // tslint:disable-line:no-any
-};
-
-export type IValidator = {
-	validate: (value: string | number | boolean | null | undefined, config: any) => false | {messageKey: string, options?: any}; // tslint:disable-line:no-any
-};
-
-export type IFormInput = {
-	name: string;
-	initialValue?: string | number | boolean | null;
-	properties: {
-		label?: string;
-		type?: 'text' | 'email' | 'password' | 'select' | 'multi-select' | 'info';
-		maxlength?: number;
-		required?: boolean;
-		showRequiredAsterisk?: boolean;
-		validators?: {
-			[name: string]: IValidator;
-		};
-		validationRules?: Array<Rule | RuleGroup>;
-		validateOnBlur?: boolean;
-		infoText?: string;
-		placeholder?: string;
-		options?: Array<{label: string; value: string}>;
-		autocomplete?: 'off' | 'new-password' | 'current-password' | 'given-name' | 'family-name' | 'email'; // https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/autocomplete
-		capitalize?: boolean;
-		focusInitially?: boolean;
-	};
-	shouldDisplay?: (values: {[key: string]: unknown}) => boolean;
-};
-
-export type IFormInputs = IFormInput[];
-
-export type IFormBoxConfig = {
-	title: string;
-	buttonText?: string;
-	secondaryButtonText?: string;
-	inputs: IFormInputs;
-	redirectLink?: string;
-	redirectText?: string;
-};
 
 export interface ITab {
 	value: string | number;
